PR Review : externalization of style functions for tabs.
