+++ src/components/dialogs/network-modifications/two-windings-transformer/modification/two-windings-transformer-modification-dialog-tabs.js
@@ -6,7 +6,7 @@
  */
 
 import { Grid, Tab, Tabs } from '@mui/material';
-import React, { useCallback } from 'react';
+import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { TwoWindingsTransformerModificationDialogTab } from './two-windings-transformer-modification-dialog';
 import { useWatch } from 'react-hook-form';
@@ -15,16 +15,7 @@ import {
     PHASE_TAP_CHANGER,
     RATIO_TAP_CHANGER,
 } from 'components/utils/field-constants';
-
-const styles = {
-    tabWithError: (theme) => ({
-        '&.Mui-selected': { color: theme.palette.error.main },
-        color: theme.palette.error.main,
-    }),
-    tabWithErrorIndicator: (theme) => ({
-        backgroundColor: theme.palette.error.main,
-    }),
-};
+import { getTabIndicatorStyle, getTabStyle } from '../../../../utils/tab-utils';
 
 const TwoWindingsTransformerModificationDialogTabs = ({
     tabIndex,
@@ -38,41 +29,29 @@ const TwoWindingsTransformerModificationDialogTabs = ({
         name: `${RATIO_TAP_CHANGER}.${ENABLED}`,
     });
 
-    const getTabIndicatorStyle = useCallback(
-        (index) =>
-            tabIndexesWithError.includes(index)
-                ? styles.tabWithErrorIndicator
-                : undefined,
-        [tabIndexesWithError]
-    );
-
-    const getTabStyle = useCallback(
-        (index) =>
-            tabIndexesWithError.includes(index)
-                ? styles.tabWithError
-                : undefined,
-        [tabIndexesWithError]
-    );
-
     return (
         <Grid container item>
             <Tabs
                 value={tabIndex}
                 variant="scrollable"
                 onChange={(event, newValue) => setTabIndex(newValue)}
-                TabIndicatorProps={{ sx: getTabIndicatorStyle(tabIndex) }}
+                TabIndicatorProps={{
+                    sx: getTabIndicatorStyle(tabIndexesWithError, tabIndex),
+                }}
             >
                 <Tab
                     label={
                         <FormattedMessage id="TwoWindingsTransformerCharacteristicsTab" />
                     }
                     sx={getTabStyle(
+                        tabIndexesWithError,
                         TwoWindingsTransformerModificationDialogTab.CHARACTERISTICS_TAB
                     )}
                 />
                 <Tab
                     label={<FormattedMessage id="LimitsTab" />}
                     sx={getTabStyle(
+                        tabIndexesWithError,
                         TwoWindingsTransformerModificationDialogTab.LIMITS_TAB
                     )}
                 />
@@ -81,6 +60,7 @@ const TwoWindingsTransformerModificationDialogTabs = ({
                         <FormattedMessage id="TwoWindingsTransformerRatioTapChangerTab" />
                     }
                     sx={getTabStyle(
+                        tabIndexesWithError,
                         TwoWindingsTransformerModificationDialogTab.RATIO_TAP_TAB
                     )}
                     disabled={!ratioTapChangerEnabledWatch}
@@ -90,6 +70,7 @@ const TwoWindingsTransformerModificationDialogTabs = ({
                         <FormattedMessage id="TwoWindingsTransformerPhaseTapChangerTab" />
                     }
                     sx={getTabStyle(
+                        tabIndexesWithError,
                         TwoWindingsTransformerModificationDialogTab.PHASE_TAP_TAB
                     )}
                     disabled={!phaseTapChangerEnabledWatch}
