+++ src/components/utils/tab-utils.ts
@@ -1,0 +1,32 @@
+/**
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+import { Theme } from '@mui/material';
+
+const styles = {
+  tabWithError: (theme: Theme) => ({
+    '&.Mui-selected': { color: theme.palette.error.main },
+    color: theme.palette.error.main,
+  }),
+  tabWithErrorIndicator: (theme: Theme) => ({
+    backgroundColor: theme.palette.error.main,
+  }),
+};
+
+export const getTabIndicatorStyle = (
+  (tabIndexesWithError: number[], index: number) =>
+    tabIndexesWithError.includes(index)
+      ? styles.tabWithErrorIndicator
+      : undefined
+);
+
+export const getTabStyle = (
+  (tabIndexesWithError: number[], index: number) =>
+    tabIndexesWithError.includes(index)
+      ? styles.tabWithError
+      : undefined
+);
